from Main

+++ app/Models/ServerLog.php
@@ -2,10 +2,15 @@
 
 namespace App\Models;
 
+use Exception;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use Symfony\Component\HttpFoundation\StreamedResponse;
+use Throwable;
 
 /**
  * @property int $server_id
@@ -15,7 +20,7 @@ use Illuminate\Support\Str;
  * @property string $disk
  * @property Server $server
  * @property ?Site $site
- * @property string $content
+ * @property bool $is_remote
  */
 class ServerLog extends AbstractModel
 {
@@ -27,11 +32,13 @@ class ServerLog extends AbstractModel
         'type',
         'name',
         'disk',
+        'is_remote',
     ];
 
     protected $casts = [
         'server_id' => 'integer',
         'site_id' => 'integer',
+        'is_remote' => 'boolean',
     ];
 
     public static function boot(): void
@@ -39,8 +46,14 @@ class ServerLog extends AbstractModel
         parent::boot();
 
         static::deleting(function (ServerLog $log) {
-            if (Storage::disk($log->disk)->exists($log->name)) {
-                Storage::disk($log->disk)->delete($log->name);
+            if ($log->is_remote) {
+                try {
+                    if (Storage::disk($log->disk)->exists($log->name)) {
+                        Storage::disk($log->disk)->delete($log->name);
+                    }
+                } catch (Exception $e) {
+                    Log::error($e->getMessage(), ['exception' => $e]);
+                }
             }
         });
     }
@@ -60,6 +73,42 @@ class ServerLog extends AbstractModel
         return $this->belongsTo(Site::class);
     }
 
+    /**
+     * @throws Throwable
+     */
+    public function download(): StreamedResponse
+    {
+        if ($this->is_remote) {
+            $tmpName = $this->server->id.'-'.strtotime('now').'-'.$this->type.'.log';
+            $tmpPath = Storage::disk('local')->path($tmpName);
+
+            $this->server->ssh()->download($tmpPath, $this->name);
+
+            dispatch(function () use ($tmpPath) {
+                if (File::exists($tmpPath)) {
+                    File::delete($tmpPath);
+                }
+            })
+                ->delay(now()->addMinutes(5))
+                ->onQueue('default');
+
+            return Storage::disk('local')->download($tmpName, str($this->name)->afterLast('/'));
+        }
+
+        return Storage::disk($this->disk)->download($this->name);
+    }
+
+    public static function getRemote($query, bool $active = true, ?Site $site = null)
+    {
+        $query->where('is_remote', $active);
+
+        if ($site) {
+            $query->where('name', 'like', $site->path.'%');
+        }
+
+        return $query;
+    }
+
     public function write($buf): void
     {
         if (Str::contains($buf, 'VITO_SSH_ERROR')) {
@@ -72,12 +121,58 @@ class ServerLog extends AbstractModel
         }
     }
 
-    public function getContentAttribute(): ?string
+    public function getContent($lines = null): ?string
     {
+        if ($this->is_remote) {
+            return $this->server->os()->tail($this->name, $lines ?? 150);
+        }
+
         if (Storage::disk($this->disk)->exists($this->name)) {
+            if ($lines) {
+                return tail(Storage::disk($this->disk)->path($this->name), $lines);
+            }
+
             return Storage::disk($this->disk)->get($this->name);
         }
 
         return '';
+    }
+
+    public static function log(Server $server, string $type, string $content, ?Site $site = null): static
+    {
+        $log = new static([
+            'server_id' => $server->id,
+            'site_id' => $site?->id,
+            'name' => $server->id.'-'.strtotime('now').'-'.$type.'.log',
+            'type' => $type,
+            'disk' => config('core.logs_disk'),
+        ]);
+        $log->save();
+        $log->write($content);
+
+        return $log;
+    }
+
+    public static function make(Server $server, string $type): ServerLog
+    {
+        return new static([
+            'server_id' => $server->id,
+            'name' => $server->id.'-'.strtotime('now').'-'.$type.'.log',
+            'type' => $type,
+            'disk' => config('core.logs_disk'),
+        ]);
+    }
+
+    public function forSite(Site|int $site): ServerLog
+    {
+        if ($site instanceof Site) {
+            $site = $site->id;
+        }
+
+        if (is_int($site)) {
+            $this->site_id = $site;
+        }
+
+        return $this;
     }
 }
